Added Host header to server-side request
This is necessary due to the way that ingress-nginx works. Ingress-nginx
can handle multiple domains inside the same cluster, for this reason,
the Host header in a request is used to specify the set of resources we
are trying to access, to when we use ingress-nginx's internal domain
address, it does not know that we actually want to request resourcces
from the ticketing.local domain. So, we must, somehow, manually add the
right Host header to the request.