Refactor successful, everything working
The new module "common" has been created, and included in the auth
service, successfully. The code that needed to be extracted from auth,
was done so without any major issue. Then the library was install in the
auth service and all relevant imports have been updated. So far,
everything working as expected.