Trying to get the desired behavior from the listener prototype
The problem is that Jetstream workqueues only accept, at least it seems,
either a single consumer or consumers with different filters (for
different topics). But, the behavior I want to achieve right now, is
intended for me to be able to horizontally scale the services and that,
at least for the particular use case I am prototyping right now,
messages are only consumed/send-to a single instance/pod of a
service/deployment, and all of them must be subscribed to the same queue
(workqueue in jetstream terms).