Listener is working, but without Event-Sourcing working
The messages are load balanced and are processed by only one subscriber,
all subscribers and bound to the same durable consumer. But, the events
are deleted from the NATS server as soon as they are consumed when using
the Workqueue setting. So, my next strategy will the to use the Limits
setting, instead of Workqueue, and use the 'subscribe' method inside the
subscribers together with the 'queue' option in the same function. This,
way, I believe, the queue will be created in the consumer/subscriber
side.