The NATS Jetstream desired behavior has been implemented
This was achieve by using the approach I was going to test, described in
the commit previous to this one.

Basically, I create a Limits JS Stream, and then, in the subscribers, I
get the messages using the 'subscribe' method, but specifying a queue,
because if this setting is missing this approach will simply not work.

The resulting behavior is that NATS persist and keeps the messages for
an unlimited period of time, because I set the Limits stream with this
configuration,  but the messages served to the subscribers are load
balanced. So, if I want to the get the full list of messages, I would
need to use/create a separate consumer that would pull all the messages
since the first one.

UPDATE: Finally found the exact behavior, using a Limits stream, and
making the scaled service instances use the same durable consumer does
the trick. By default, messages for subscriptions using the same durable
consumer are "load balanced" (at least distributed) between
instances/subscriptions, thus achieving the desired effect.

This is accomplished by using the code in publisher_limits.ts as is,
with the consumer configuration in listen.ts as is.