Solved the issue with the Tickets service
As I am using minikube with Skaffold, and set the policy to
never-to-pull, Skaffold was using the images stored localy in minikube's
docker server. This was expected, the issue was, that minikube/Skaffold
was only re-building the upper layers of my Tickets service, after
library installs. Meaning that the new version of my common library were
not being installed.

The simple solution was to manually update the minimum required version
for this packaged in the package.json file. This triggered a rebuild
previous to the library-installation step.