Added functionality to test whether the stream already exists
More accurately, test if the stream has already been configured on the
server. If the stream already exists, the code assumes that the stream
has been also configured with the appropriate subjects.

As a side note, it might, or might not, be a good idea to let services
register their own events by means of editing the nats stream
configuration. Let it be noted that doing this might not be compatible
with other message-passing software.