Modified the natsWrapper mock
Now, it makes use of jest.fn function in order to track calls to the
publish function. As it is now, the code will not work. So, I will have
to do some refactoring until jest can track the calls. For this, I will
need to change from passing the NATS connection to passing the client
object around.