First attempt in correctly extending nats-wrapper's mock
I have created a mock for the attribute/function 'connection' from
natsWrapper. I have made it so that the connection  so that the
jetstream function now returns null. This was done because I think it
will not affect the rest of the test.

Other tests, like the new.test.ts, still present errors. It seems they
do not use the id property of the tickets object when building. Got to
check that out next.