OrderCreatedListener now publish the TicketUpdatedEvent
This was done so the copy of the version attribute in other services,
introduced for OCC, does not get out of sync with the version
in the service in charge of this resource, thus failing further updates
to the services listening for changes to the Tickets resources.

Also, updated the tests to include a test for this functionality.