Added K8s config for stripe's webhook secret
Also modified the apiVersion line in stripe.ts. This was needed because
of a mismatch between the string literal VSCode's linter expected and
the one that the compiler expected. The new string literal is what the
compiler expects to encounter.

A future improvement would be to store this api version string inside an
environment variable.