Added server logging of webhook errors
This has to be done, because the response (and error information) that
the "res" object sends is, in localtest, sent to the stripe mock command
sending the fake requets to the webook, while in production, this is
send to the stripe service.

In both cases, this implies that I would need to check the stripe
dashboard for information on the error instead of my own logs. As this
is unacceptable, I add the server logging of such events.

One caveat is that while testing locally I can get the error messages in
the responses to the stripe client using the following command (with the
option at the end as a must):
    stripe trigger payment_intent.succeeded --log-level debug