Fixed profile mixup
the activation, as it was set, made that profile run everything the
command 'dev' was run. What this did, was in effect introducing the same
bug we were trying to prevent by separating profile.

Now, there is not a 'default' profile. Each time I run the 'skaffold
dev' command, I must assign a specific profile name.